pyproject.toml
[project]
name = "thrive-well-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-sugar>=1.0.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]
